Added new hook for covering executor and task environment.
For being able to supply environment additions for both, a task and
its executor separately we need to introduce a new hook as the
existing ones slavePreLaunchDockerHook (deprecated) as well as
slavePreLaunchDockerEnvironmentDecorator do not allow for this.

This new hook will likely allow for further additions like e.g.
adding volumes without having to adapt the signature but only the
returned proto message TaskExecutorDecoratorInfo.

Review: https://reviews.apache.org/r/54038/