Stopped sending checkpoint resources message on agent re-registration.
Given that resource provider capable agents will send update slave
message to the master during re-registration, no need for the master
to send checkpoint resources message to the agent anymore.

This also makes the code more consistent because agent should be the
source of truth. This also eliminates the possible retry incurred by
this message, which is never the intention.

Review: https://reviews.apache.org/r/62879