Added a mock resource provider manager.
This patch adds a `ResourceProviderManager*` parameter to the slave
constructor, so we can pass in a mock resource provider manager. The
mock manager can be used to test against either the agent or the
resource provider.

The declaration of `ResourceProviderManagerProcess` is moved to a
separated `manager_process.hpp` header file so we can mock the internal
process.

Review: https://reviews.apache.org/r/63389