Allowed relative path in `Resource.disk.source.{path,mount}.root.
If the `root` field is set to a relative path, it would mean that the
path is relative to the agent work directory. This provides the ability
for an isolator to construct the mount point of a CSI volume without
knowing the resource provider info.

Review: https://reviews.apache.org/r/63498