Added a container daemon to monitor a long-running standalone container.
The `ContanierDaemon` class is responsible to monitor if a long-running
service running in a standalone container is ready to serve, and restart
the service container if not. It takes the following 3 hook functions:

* `preStartHook`: called prior to container launch.
* `readinessHook`: called periodically to check if the service is ready.
  If a duration is returned, it serves as a backoff period, meaning that
  the next check will be scheduled after the duration. It can also be
  used to set up futures to reflect the readiness of the status outside
  the container daemon.
* `postStopHook`: called after the container is killed.

`ContainerDaemon` does not manage the lifecycle of the contanier it
monitors, so the container persists across `ContainerDaemon`s. However,
it provides a `terminate()` method to completely shutdown the monitoring
container.

Review: https://reviews.apache.org/r/63680