Added `getService()` function to launch CSI plugins.
The `getService()` method first checks if there is already a container
daemon for the specified plugin, and creates a new one if not. The
post-start hook for the container daemon will call `connect()` to
wait for the endpoint socket file to appear and connect to it, then
set up the corresponding promise of CSI client. The post-stop hook will
remove the socket file and create a new promise for the next call to the
post-start hook to set it up.

Review: https://reviews.apache.org/r/63021/