Fixed composing containerizer hanging in `wait()` after recovery.
This patch fixes the issue when `wait()` of composing containerizer
returns a future that is never set to `READY` state due to a missing
`wait()` call on corresponding containerizer after it's recovery.
It is necessary to call `wait()` for each known container of composing
containerizer, because any terminated container should be handled in
`_destroy()` callback that is bound to that `wait()`.

Review: https://reviews.apache.org/r/65071/