Introduced a push-based gauge metric.
A push-based gauge differs from a pull-based gauge in that the
client is responsible for pushing the latest value into the gauge
whenever it changes. This can be challenging in some cases as it
requires the client to have a good handle on when the gauge value
changes (rather than just computing the current value when asked).

It is highly recommended to use push-based gauges if possible as
they provide significant performance benefits over pull-based
gauges. Pull-based gauge suffer from delays getting processed on
the event queue of a `Process`, as well as incur computation cost
on the `Process` each time the metrics are collected. Push-based
gauges, on the other hand, incur no cost to the owning `Process`
when metrics are collected, and instead incur a trivial cost when
the `Process` pushes new values in.

Review: https://reviews.apache.org/r/66828/