Fixed installation of storage disk profile adaptor header.
Automake does not care about the source directory of headers it
installs, but instead either installs into `$(pkgincludedir)` or some
`COMPONENTdir`. When doing in-tree builds on the other hand the
header source directories matter since the build is performed against
a source tree, not an installed tree.

This patch fixes installation of the `disk_profile_adaptor.hpp` header
which while it resided in the source directory
`mesos/resource_provider/storage` was installed into
`mesos/resource_provider`. This made e.g., other headers relying on
the disk profile adaptor header unusable when building against
installed Mesos headers. We now make sure to install to the `storage/`
subdirectory.