Added missing iptables argument in CNI port mapper.
Previously, the CNI port mapper was using the `-w` option in all
invocations of `iptables` except one. This option ensures that if
an application is currently holding the xtables lock, then `iptables`
will wait for the lock to be released and then continue execution.
Without setting the `-w` option, `iptables` will fail immediately if
the lock is currently held.

This patch adds the missing `-w` option to the port mapper code.

Review: https://reviews.apache.org/r/68152/