Updated handleWhitelistFds() to avoid closing FDs with FD_CLOEXEC bit.
Since this helper is only called right before exec in child process,
and for those open FDs that is set with FD_CLOEXEC flag, they will
be closed during exec, so that we could skip closing these FDs in
the helper. The motivation of this change is to avoid whitelisting
those FDs that have to survive until exec while we do not want to
expose these FDs to user applications.