Added a registry operation for the DRAIN_AGENT call.
This adds the associated registry operation and fields for the
DRAIN_AGENT master call.  This call affects admitted or unreachable
agents, but this commit only deals with admitted agents.

Because this feature is not downgrade compatible, a minimum capability
is added when the draining feature is used.

Review: https://reviews.apache.org/r/70923/