Implemented master endpoints for agent draining.
This fleshes out three master calls:
  * DRAIN_AGENT
  * DEACTIVATE_AGENT
  * REACTIVATE_AGENT

The master now stores a mapping of agent draining or deactivation
information.  When an agent reconnects, this information is checked
before informing the allocator about the agent.

This commit leaves out certain aspects of each endpoint's validation,
such as checking if draining agents are not in maintenance schedules.

The DRAIN_AGENT call is not completely implemented here, because the
transition from DRAINING to DRAINED will be added in a separate commit.

Review: https://reviews.apache.org/r/70996/