Implemented transition from DRAINING to DRAINED in master.
This adds logic in the master to detect when a DRAINING agent can
be transitioned into a DRAINED state.  When this happens, the new
state is checkpointed into the registry and, if the agent is to be
marked "gone", the master will remove the agent.

Review: https://reviews.apache.org/r/71008/