Disabled hostname validation for master to scheduler communication.
Some schedulers may run in non-host networking mode containers, in which
case the default-provisioned certificate from the DC/OS CA does not
contain a corresponding subject alternative name, making hostname
validation fail.

In DC/OS, this might happen e.g. when a spark scheduler is launched
with the setting `virtual_network_enabled:true`.

As a workaround, we enforce the hostname "localhost" which is contained
in all certificates generated by the DC/OS CA.